+++ lib/dogen.rb
@@ -2,12 +2,12 @@
 
 require_relative "dogen/version"
 require_relative "dogen/arguard"
-require_relative "dogen/render"
 require_relative "dogen/decorator"
+require_relative "dogen/generator"
 require_relative "dogen/branded"
-require_relative "dogen/dom"
+require_relative "dogen/domain"
 require_relative "dogen/dsl"
-require_relative "dogen/gen"
+require_relative "dogen/cli"
 
 module Dogen
   class Error < StandardError; end