+++ test/dogen/domain_guards_spec.rb
@@ -1,5 +1,21 @@
 require_relative '../spec_helper'
-require_relative 'shared_guard_specs'
+
+module SharedGuardSpecs
+  extend Minitest::Spec::DSL
+
+  # must be provided in target module:
+  #   let(:guard) { GuardName }
+  #   let(:valid) { ["name", :name] }
+  #   let(:wrong) { [nil, 1, Object.new]}
+
+  it 'must return value' do
+    valid.each{|v| assert_equal v, guard.(v)}
+  end
+
+  it 'must raise ArgumentError' do
+    wrong.each{|w| assert_raises(ArgumentError) { guard.(w) }}
+  end
+end
 
 describe 'Domain Guards' do