+++ test/dogen/dsl_spec.rb
@@ -1,10 +1,10 @@
 require_relative '../spec_helper'
-require_relative 'shared_domain'
+require_relative 'demo_domain'
 
 describe DSL do
 
   it 'must self#build complex domain' do
-    assert_instance_of Domain, build_users_domain
+    assert_instance_of Domain, build_demo_domain
   end
 
   it 'DSL must be extensively tested' do